use libpcap 1.5
As it turns out, tcp(1) broke with libpcap 1.5 because of breaking changes
in the API, documented in pcap_set_immediate_mode(3pcap).  It still worked,
kind of, but live packet input got delivered in batches, and you wouldn't
see any output until a sizeable batch of frames was pending.

So I'm moving to the new API so that I can call the critical new
pcap_set_immediate_mode().  Hopefully it's available everywhere that
I care about.

The recent libpcap maintainers don't seem that bright. See this writeup
for a partial analysis and seemingly initiated complaints:

  <https://blog.erratasec.com/2018/07/some-changes-in-how-libpcap-works-you.html>